Update index.js
Try to stop mouseover, mouseout events after click any star by if/else conditions.  I am close to the end of the task, but this is still not perfect.

Rewriting symbols by reset button not working as it is should - level is not returning to 'easy level' (level = 1), elements are still at the same level as the last time (for example 3).
Hard reset (refresh all website - as F5 key press), which I try to make (line 182-183).
Another idea was written between line 178-181 (makes a function with some variable=1 and recall newSymbols( ) ) . All with no success.